Fixed bugs

+++ CppND-System-Monitor-Project/src/processor.cpp
@@ -32,8 +32,8 @@ float Processor::Utilization() {
   long int totald = Total - PrevTotal;
   long int idled = Idle - PrevIdle;
 
-  long int CPU_Percentage = (float)(totald - idled)/(float)totald;
-    
+  float CPU_Percentage = (float)(totald - idled)/(float)totald;
+
   prevuser = user;
   prevnice = nice;
   prevsystem = system;
